Check for HMAC-MD5 instead of just MD5
MD5 is often used as a non-cryptographically secure CRC function. And thus
availability of MD5 is often fine.

What matters the most is that cryptographic functionality with MD5 is blocked,
such as HMAC, symmetric signature creation or verficiation, key derivation,
etc.

HMAC is the simpliest of them all, thus switch the unapproved algorithms check
to check for HMAC-MD5. This way even if MD5 digest is allowed, it is clear that
cryptographic protection with MD5 is blocked.